MAINT: Move TimeDependent, Iterative and Reference operators to separate module
Also make these mixins to be added to operators
