+++ src/porepy/numerics/ad/surrogate_operator.py
@@ -131,7 +131,8 @@ import porepy as pp
 from porepy.numerics.ad.forward_mode import AdArray
 
 from .functions import FloatType
-from .operators import IterativeOperator, Operations, TimeDependentOperator
+from .operators import Operator, Operations
+from ._derived_operators import TimeDependentOperator, IterativeOperator
 
 __all__ = [
     "SurrogateOperator",
@@ -139,10 +140,7 @@ __all__ = [
 ]
 
 
-class SurrogateOperator(
-    TimeDependentOperator,
-    IterativeOperator,
-):
+class SurrogateOperator(TimeDependentOperator, IterativeOperator, Operator):
     """Operator representing a :class:`SurrogateFactory` in AD operator form on
     specified subdomains or interfaces, at a time or iterate index.
